+++ plugins/postcss-custom-properties/src/lib/get-custom-properties-from-root.ts
@@ -2,10 +2,11 @@ import valuesParser from 'postcss-value-parser';
 import { isBlockIgnored } from './is-ignored';
 
 // return custom selectors from the css root, conditionally removing them
-export default function getCustomPropertiesFromRoot(root, opts) {
+export default function getCustomPropertiesFromRoot(root, opts): Map<string, valuesParser.ParsedValue> {
 	// initialize custom selectors
-	const customPropertiesFromHtmlElement = {};
-	const customPropertiesFromRootPseudo = {};
+	const customPropertiesFromHtmlElement: Map<string, valuesParser.ParsedValue> = new Map();
+	const customPropertiesFromRootPseudo: Map<string, valuesParser.ParsedValue> = new Map();
+	const out: Map<string, valuesParser.ParsedValue> = new Map();
 
 	// for each html or :root rule
 	root.nodes.slice().forEach(rule => {
@@ -18,11 +19,11 @@ export default function getCustomPropertiesFromRoot(root, opts) {
 		// for each custom property
 		if (customPropertiesObject) {
 			rule.nodes.slice().forEach(decl => {
-				if (isCustomDecl(decl) && !isBlockIgnored(decl)) {
+				if (decl.variable && !isBlockIgnored(decl)) {
 					const { prop } = decl;
 
 					// write the parsed value to the custom property
-					customPropertiesObject[prop] = valuesParser(decl.value);
+					customPropertiesObject.set(prop, valuesParser(decl.value));
 
 					// conditionally remove the custom property declaration
 					if (!opts.preserve) {
@@ -38,21 +39,25 @@ export default function getCustomPropertiesFromRoot(root, opts) {
 		}
 	});
 
+	for (const [name, value] of customPropertiesFromHtmlElement.entries()) {
+		out.set(name, value);
+	}
+
+	for (const [name, value] of customPropertiesFromRootPseudo.entries()) {
+		out.set(name, value);
+	}
+
 	// return all custom properties, preferring :root properties over html properties
-	return { ...customPropertiesFromHtmlElement, ...customPropertiesFromRootPseudo };
+	return out;
 }
 
 // match html and :root rules
 const htmlSelectorRegExp = /^html$/i;
 const rootSelectorRegExp = /^:root$/i;
-const customPropertyRegExp = /^--[A-z][\w-]*$/;
 
 // whether the node is an html or :root rule
 const isHtmlRule = node => node.type === 'rule' && node.selector.split(',').some(item => htmlSelectorRegExp.test(item)) && Object(node.nodes).length;
 const isRootRule = node => node.type === 'rule' && node.selector.split(',').some(item => rootSelectorRegExp.test(item)) && Object(node.nodes).length;
-
-// whether the node is an custom property
-const isCustomDecl = node => node.type === 'decl' && customPropertyRegExp.test(node.prop);
 
 // whether the node is a parent without children
 const isEmptyParent = node => Object(node.nodes).length === 0;